cache/ttl: move ttl files to its own folder
