+++ cache/ttl/ttl.go
@@ -1,4 +1,4 @@
-package cache
+package ttl
 
 // import (
 // 	"sync"
