+++ cache/ttl/ttl_suite_test.go
@@ -1,0 +1,13 @@
+package ttl_test
+
+import (
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+
+	"testing"
+)
+
+func TestTtl(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Ttl Suite")
+}
